complete all three crud with schema

+++ app/routers/user_router.py
@@ -1,7 +1,7 @@
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, HTTPException
 from app.config.config import SessionLocal
-from app.schemas.schema import UserSchema
-from app.services.user_crud import create_new_user, delete_user, get_all_user, get_user_by_id, update_user
+from app.schemas.schema import UserCreateSchema, UserSchema, UserUpdateSchema
+from app.services.user_crud import create_new_user, delete_all_user, delete_user, get_all_user, get_user_by_id, update_user
 from sqlalchemy.orm import Session
 
 from app.models.models import User
@@ -21,10 +21,10 @@ async def get_users(db:Session=Depends(get_db)):
     return all_user
 
 
-@user_router.post("/create")
-async def create_user(request: UserSchema, db:Session = Depends(get_db)):
-    new_user = create_new_user(db, request)
-    return {"message":"user created", "user":new_user}
+@user_router.post("/create/")
+async def create_user(user: UserCreateSchema, db:Session = Depends(get_db)):
+    new_user = create_new_user(db, user)
+    return {"user":new_user,"message":"user created", }
 
 
 @user_router.get("/user/{id}")
@@ -34,13 +34,28 @@ async def user_by_id(id:int, db:Session = Depends(get_db)):
 
 
 @user_router.put("/update/{id}")
-async def updat_user(id:int, user:UserSchema, db: Session = Depends(get_db)):
-    updated_user = update_user(user.id, db, user.name, user.age)
-    return updated_user
+async def updat_user(id:int, user:UserUpdateSchema, db: Session = Depends(get_db)):
+    existing_user = db.query(User).filter(User.id == id).first()
+
+    if not existing_user:
+        raise HTTPException(status_code=404, detail="Entry not found")
+    else:
+        updated_user = update_user(id, db, user.name, user.age)
+        return updated_user
 
 
 @user_router.delete("/delete/{id}")
 async def delete_user_(id:int, db:Session = Depends(get_db)):
-    delete_user(id,db)
-    return {"message":"User deleted"}
+    existing_user = db.query(User).filter(User.id == id).first()
 
+    if not existing_user:
+        raise HTTPException(status_code=404, detail="Entry not found")
+    else:
+        delete_user(id,db)
+        return {"message":"User deleted"}
+    
+
+@user_router.delete("/delete_all/")
+async def delete_all(db:Session = Depends(get_db)):
+    delete_all_user(db)
+    return {"message":"Delete all users"}
